Working on clean version with updated math
